Fix compatibility with Django < 1.4.3
The is_safe_url function used in a previous commit was introduced in
Django 1.4.3. I think breaking compatibility with the old version is
unnecessary, and Ubuntu has backported the security fixes so I'm
hesitant to require a new Django version.

This commit moves the function into openstack_auth.utils, and uses
it if the Django version is missing.