Remove admin role name 'admin' hardcode in User.is_superuser()
Because of hardcoding name as the 'admin' was impossible to
use administrative panel with a custom administrative role name.
This fix replaces hardcoded the name of the administrative role
with the Horizon settings constant with list of roles.
So, now administrative roles can be multiple.
Related commit: https://review.openstack.org/#/c/123741/

Change-Id: I04277cd50a938f02949ae0e33228e1628197d252
Partial-Bug: #1161144