Add missing olso library entries to requirements.txt
Some oslo libraries are used indirectly via the sync of
oslo-incubator but were not included in the requirements.txt.
This caused them to not be capped and to pick up versions of
the libraries that no longer support the old namespace.

The libraries are:
    oslo.config
    oslo.i18n
    oslo.serialization
    oslo.utils

Note this isn't a problem in master because it has moved to
using the library versions of the oslo-incubator support that
was being used.

Change-Id: Ia53b59b9a1ea3440ed3c335bef4c69e44a53df0f
Closes-Bug: 1470574