Replace default User model PK
The default Django User model PK is an int() AutoField
and django-openstack-auth sets this to a hash string. Django
then breaks trying to coerce that string to an int().

This patch adds a new explicit PK to the d-o-a User
model. It also adds the standard Django "models.py" so
that the consumer application (Horizon) may use it.

The consumer application must set:

   AUTH_USER_MODEL = 'openstack_auth.User'

to use the new model in place of the default 'auth.User'.

The approach in this patch was inspired by Lin Hua
Cheng <os.lcheng@gmail.com>.

Partial-Bug: 1491117
Change-Id: I549209eb0bb0ddf36d92ee9dc1a9bac799ce67e5