Unscoped PKI token should no longer be hashed multiple times.
When token hashing is used with pki tokens, the unscoped token
gets re-hashed when switching project. This fix checks
if the unscoped token needs to be hashed before doing so.

The project list operation when switching project in horizon
failed because the unscoped token could for example be an
md5 of an md5.

Change-Id: I64684ca251eb4d0c6164e58c078cf7d132eb3cc1
Closes-Bug: #1487372
(cherry picked from commit 5ab3908cff64fa5b5a5bd5ea4877c42096172dd3)