Change log.error to log.warning
Looks like the actual repercussion from this "error" is
nothing at all, so the log should be changed to a warning
as _domain_to_credentials returns even if an exception is
raised.

So when a function like this works even if an exception is
raised, the log level should be a warning at best, as the rest
of the code depending on this still seems to work no matter
what is returned.

Change-Id: I8ea690923010674351d45ef5122bb046ee5e46b5
Closes-Bug: #1527575