Add Constraints support
Adding constraints support to libraries is slightly more complex than
services as the libraries themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove django_openstack_auth.

Change-Id: Ibb739fe4385caa3dd9f96f0708b7b3830a915d07