import zuul job settings from project-config
This is a mechanically generated patch to complete step 1 of moving
the zuul job settings out of project-config and into each project
repository.

Because there will be a separate patch on each branch, the branch
specifiers for branch-specific jobs have been removed.

Because this patch is generated by a script, there may be some
cosmetic changes to the layout of the YAML file(s) as the contents are
normalized.

See the python3-first goal document for details:
https://governance.openstack.org/tc/goals/stein/python3-first.html

* For docs job
The older branches did not work with the existing documentation job.
The fix may require backporting changes to remove tox_install.sh, or
other changes that have been made in newer stable branches but not
backported all the way. Because the new documentation job runs through
tox we may be able to use that in the older branches, as an alternative.

Change-Id: I311741b1b7b2617e304b7fb4568e2a9b1d3238b1
Story: #2002586
Task: #24299